Disable Turbo Links in all Forms and Links of Devise Views
note: At the time of writing, Devise and Turbo Drive don’t play nicely together. When using devise with Turbo Drive you have two options. Either you can generate the devise views to your local app (covered in the devise README) and then for each view with a form disable turbo drive by adding the data attribute. This is time consuming but simple to follow. The other way is to create a custom controller to handle this and then use devise for that. This GoRails episode covers how you’d do that. You might not understand everything being done but it’s a quicker solution. Hopefully Devise will be Rails 7 Turbo compatible by the time you read this